backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
botocore==1.34.122
certifi==2024.2.2
chalice==1.31.2
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
frozenlist==1.4.1
idna==3.7
inquirer==2.10.1
itsdangerous==2.2.0
jmespath==1.0.1
MarkupSafe==2.1.5
multidict==6.0.5
newsapi-python==0.2.7
packaging==24.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
PyYAML==6.0.1
readchar==4.1.0
regex==2024.5.15
requests==2.32.2
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
